added about.html and about.js,refactored scripts
Added an about page that tells you about how the website is meant to be used and it also display your currently tracked champions. Made scripts.js more readable with comments and some refactoring.  todo: make the champion boards prettier and make the website more accesible. Finetune the searchbar a bit more so it feels better to use, currently some small bugs in some cases but works fine for now